Refactor stubbing of jclouds for tests
